Automatically switch protocol versions to find the best one
There are two protocol versions that Goldair devices may use. Rather
than requiring that the user know which one and put it in the config, a
more automatic solution is to just switch the version when there is a
communication error. There may be some false positives when network
connectivity is low, but it's a much nicer solution than putting it on
the user.