Add support for GPCV and GECO series heaters.
These are different than the GPPH series heaters that are already supported.
See Issue #27

Based on changes in KiLLeRRat/homeassistant-goldair-climate and the manual for GECO/GPEH heaters, the DPS arrangement for GECO seems the same as GPCV, but this model does not have the High and Low modes which we exposed as PRESET_MODE in the GPCV so may generate errors if configured as GPCV.

This heater appears cosmetically and functionally the same as GEPH, so hopefully it works with these as well.