Swap symlinks around to avoid HACS breakage.
Symlinks are trying to support different versions of HomeAssistant.
Latest 0.109 uses translations/en.json, so make that the master copy of
the file, and make strings.json a symlink to that.
Update hacs.json to indicate 0.109 is the minimum version - although the
source code supports 0.96 and later, the way HACS breaks the symlinks on
installation means it will only work on 0.109 if installed through HACS.