Add support for GECO Heaters
Speculative based on changes in KiLLeRRat/homeassistant-goldair-climate and the manual for this model.
The DPS arrangement seems the same as GPCV, but this model does not have the High and Low modes which we exposed as PRESET_MODE in the GPCV so may generate errors if configured as GPCV.

This heater appears cosmetically and functionally the same as GEPH, so hopefully it works with these as well.