Be less eager about rotating the protocol version once it is working
Sometimes devices fail to respond, or respond with a response that pytuya
cannot parse.  This happens especially with the encrypted 3.3 protocol.
Rotating the protocol version seems to make the problem worse, causing the
device to stop responding for some time.
Let the retry counter reach _CONNECTION_ATTEMPTS before starting to rotate
the protocol version if we have successfully communicated until then.