Use lock when getting status from device.
Sending commands to the device was protected by a lock, but getting status
could still run in parallel.
This can cause the device to fail to respond, as the Tuya firmware only seems to handle one connection at a time.
Use the same lock so that status requests will wait for commands before firing.