Rename some fields for consistency.
primary_device renamed to primary_entity, since in HA device refers to the whole device, and each functional part is an entity.

secondary_devices renamed to secondary_entities for the same reason.

value under conditions changed to dps_val to be consistent with mappings. The values for comparison are also changed to be raw DPS values before any mapping.

value-override under conditions changed to value.  After renaming the above, there is no conflict, so the naming and meaning can be consistent with mappings.