Switch out the underlying library to tinytuya.
pytuya is unmaintained, and there seem to be several improved versions since by Tradeface (also seems to be abandoned) and James Cox (tinytuya) which still seems to be actively maintained.

Hopefully this will improve reliability, especially with the 3.3 protocol version, which still seems flakey with pytuya.  We may need to look at using some of the higher level methods to take advantage of some of the improvements, as the current Device implementation works with some underscore methods that were probably intended to be internal.