Fix teardown error due to a library update.
After updating some pytest related libraries to test another component,
I started to get teardown errors due to a Thread left running.
Since device.py contains a one second debounce Timer on network operations,
with a faked immediate set of the property, when we test this, we need to
wait for the Timer to expire to avoid the teardown error.