Improve test coverage for generic components.
Still not 100%, as the components cover some cases that our current config files don't use.  Probably some test configs are needed to fully exercise the generic classes ahead of any real device using them.

Remove devcontainer config.  Based on upstream changes, it seems it is out of date, and I do not use it myself so it is just a maintenance burden keeping it.