Implement a basic generic climate entity.
Functionality is quite limited at first, but it should cover most of the existing and pending heaters.

Implement ranges in TuyaDpsConfig, and fix the logging in map implementation.
Use the TuyaDpsConfig getter and setters in lock and switch entities (light was already accidentally committed along with another bugfix).

Remove all the simple heaters.

The generic climate class can handle the following heaters, so their specific classes can be removed:
- eurom_600_heater
- geco_heater
- gpcv_heater
- gsh_heater
- kogan_heater