Output retry log message as debug instead of warning.
Although it is a communication failure, it is expected when the protocol
version has not been detected yet. For persistent failures there is an error
log message after the retry limit is reached.

Include the device name in most log messages, making it easier to see what is
happening.

Don't count updated_at entry when judging match quality.

dps from tinytuya api has an updated_at entry in addition to actual dps.
Don't let this lower the match quality.
This should move everything from an 83.3-92.85% match to a 100% match.