GPPH heater config: add values in map even though they are the same as dps_val.
This list of dps_val was supposed to indicate that the dps is an enum, with a list
of fixed values to choose from.
When handling these, we actually check values, not dps_val.  But if we leave off the dps_val, they all become defaults, which will override any dps_val.
To avoid overcomplicating the handling code, include values for these.