Add Eanons Humidifier as a climate device.
Eanons Humidifier is added as a climate device (since that is what is currently supported, and like the Goldair Dehumidifier, this seems to have fan mode control which is missing from the humidifier entity in Home Assistant.

But with both devices, the fan mode appears to be coupled to presets, and can be covered by presets, so in future I will implement a generic humidifier entity and change both devices to that (will also require a generic sensor entity to surface the temperature sensor on the Goldair dehumidifier more usefully than as an attribute).