Allow dps to be duplicated in the config file.
This allows for additional sensor to be defined in addition to just
including them as attributes for convenience, or alternate types such
as a humidifier offering a humidifier and a climate option depending
on user preference. The latter has more functions and is better
supported in the front end, even though it does not really fit
humidifiers, but the former is the modern way of doing things even
though the front end is not really supporting it well yet.