Add a generic humidifier entity.
Humidifiers and dehumidifiers are supposed to be represented in HomeAssistant by
the humidifier entity type, available since 0.102.

Add an implementation of this, and add config for Goldair Dehumidifier and Eanons Humidifier as a secondary device for testing.  Later this will be made
the primary device (or maybe I'll get rid of the concept of primary and secondary altogether).  I'm not sure what happens if both are configured together. If there are no ill effects, I'll probably leave both configurations as it comes down to personal choice which to use.