Deprecate YAML configuration.
Soon work will start on improving the config flow.  I do not want to be
held back by legacy config format issues, so deprecating YAML now.
It is already imported since before the start of the fork, but currently
does not allow reconfiguration of imported config, and picks up changes
from yaml config.

Next step will be to ignore yaml config, and allow previously imported
config to be modified from the options flow.  After that, migrations will
not be blocked by the need to keep importing the legacy yaml config.