Resolve CONF_TYPE_AUTO once per device.
During init, when device type is set to auto, the device needs to be polled once before the device type can be resolved.  But when creating multiple entities per device, the reply from the earlier entities still hasn't come back, so we may end up creating multiple refresh requests.
Resolve the CONF_TYPE before setting up platforms so that the platforms can
be setup more quickly.

Also, reverse the logic for deciding when to rotate the protocol version, so that devices that have never communicated rotate immediately for faster detection.