Ensure cached state is fetched before use.
async_inferred_type was getting the cached_state, then looping over
async_possible_types, relying on that to ensure the cached_state was fetched
from the device.  Then it was using its own copy of cached_state to make
the decision about match quality.  But if that copy was taken before the
state is fetched from the device, it is empty.  Defer getting the copy of
cached state until inside the loop.

Improve migration logic, return the migration status.

Update strings for Options flow, even though we haven't made the changes for that yet.