Handle conditional re-mapping of values.
Use it instead of complex mix of range and mapping in fan climate config.
Only one test remains skipped in fan, which is that fan mode cannot be set when no preset is set. But since that doesn't happen in reality, we can probably switch to using the generic climate class for fan now, leaving only dehumidifier and gpph heater legacy classes.