Auto detect stringified dps.
Remove the explicit stringify property from the config files.
Instead, detect when reading the property whether it is quoted, and if so,
quote it when writing too.

This compensates for the fact that previously added devices did not pay much
attention to this, and future software updates may change the use of quotes
for those devices that do come marked as stringified.