Identify the integration consistently as "Tuya Local".
The different wording in HACS, Integrations UI and github project name are
likely to be confusing, and may have already been responsible for Issue #21.

Although it is also confusing when another unrelated popular integration
is called Local Tuya, it seems like it will be disruptive to change the
actual integration name.  But maybe migration will help, so I may look into
that in future too.

Also bumped the version in the manifest, which has been forgotten the past
few versions.