Rename poolex_heatpump to poolex_silverline_heatpump.
Per the last commit by lperez31, there are incompatible models from the
same manufacturer that need to be differentiated.

As this has not gone into a release, there is no migration for this rename.
Anyone following the main branch and using this device will need to delete it
and set it up again.