Refactor tests for devices.
Remove a lot of duplication for general device tests by using a base class.

This should reduce the cut and paste effort to add a new device with unit
tests.  Only tests for actual configured functionality should be required,
and the setUp is mostly implemented in the base class so there is no need
to mess with patching in every device specific test module.