Migrate legacy types to config filename based types.
It is more efficient to load config files based on the type, rather than having
to iterate through all configs looking for the matching legacy_type.

Change legacy type in smartplugv2 to kogan_switch2 to tie it back to the old
config name.  kogan_switch has special handling in the migration to handle
mapping to the old shared legacy_type.