Beca thermostat: split into C and F variants.
The device is switchable between C and F, but the local protocol does
not return any info about which units are in use.  Require the user to
select the correct config.
Added ability to set the units in config by putting a unit attribute on
the temperature dps, use it in the new configs.