Remap scaled and stepped values
If a mapping causes a value to be scaled or stepped due to a default mapping,
check the map again for a matching rule.

Use case:  For a light that supports 4 brightness levels that are mapped to
strings: level0, level1, level2, level3, these can be mapped to values
0, 85, 170, 255 for the standard 0-255 brightness range.  But if a different
brightness is passed in, we want it to snap to one of those values
- a step of 85 as the default mapping should do that, but we need to pass
it back through the mapping after snapping to the step to translate to the
strings.