Siswell thermostat: use separate c and f temperature dps to set temperature.
This device appears to have 3 dps corresponding to the set temperature, dps 2 which is in tenths of degrees, dps 114 which is always C, in degrees, and dps 116 which is always F, in degrees.

Previously we tried to use dps 2, as it has more precision, but this does not seem to work (comment on Issue #45).  Try the others instead.