Saswell Thermometer: rearrange temperature and mode DPS.
Based on new information from Issue #45, the mode dps is 4 not 103, so swap hvac_mode and hvac_action.

Also, the temperature setting is confirmed to be working from dps 2, so use that as it allows more precision.  Assume 0.5 step for C, and 1 for F, which seems fairly standard.  To avoid issues with temperature_unit not being as expected, default to C, and only define F as the exception.