Prepare initialisation for future multiple entities of the same type.
Index entities in the HA data store by their config_id instead of hardcoded strings.
Loop through all entities instead of breaking when the first is found.
Currently if there are multiple entities of the same type defined in a config,
this will create them, but override with subsequent ones (orphaning the earlier one, as it is not destroyed).  Further changes are needed before it is safe to create configs with multiple entries of the same type, but this change is being committed first to ensure the current configs are not affected before making the breaking config change and migration to new config_ids.