Add support for Eberg Qubo Q40HD portable air conditioner
Issue #35

Modes guessed based on other models that use "hot" as a mode and the modes listed in the manual.  Manual lacks details of the "wind" mode, does not list it with the modes but says that the device can be used as an air filter without any heat/cool/dry functionality, so assume it is there.
Also guessed that other fan modes are called "low", "high" and "auto" based on the known value of "medium".  Some other devices that use "medium" use "big" and "small" (direct translations from Chinese), but those devices are all dehumidifiers so it is not a safe assumption that there will be some common source code base with this device.
One of the unknown booleans is likely the Sleep function, and dps 101 is likely usable as hvac_action, but more details are required for that.