Add ability for climate device to specify min and max temperature.
Some devices feedback the minimum and maximum temperature range for the mode
they are in on dps.  Pass this through to the UI so the correct range is
displayed even if the logic is too complex to model in the config.
Service call failure will still depend on the range specified in the
temperature dps, but at least the UI won't have to allow setting outside the
range to avoid over-restricting.