Resolve skipped tests for GPPH heater and remove special class for handling it.
Cross interaction between power level and "swing mode" can be represented in the config.  Remaining skipped test of restricting values does not match with the
way that we allow unknown values to go straight through without mapping, but the list of possible values is still taken from the mapping list.