Remove remaining use of legacy_class.
Dehumidifier climate entity is deprecated and the dehumidifier is working as humidifier and fan entities without the features that aren't yet implemented.
So its removal along with test cases that cannot be covered by the generic climate implementation is not much of a loss.

Remove the now unused legacy_class functionality.