Add  timer function to simple switch.
I checked 4-5 brand of simple plug (without led or measuring) and it almost all only contain this two dp_id. I think this config help to use this switches all available local function. But also usefull that the original simple switch config is also remain, so i add it to a different name.