Reduce debounce on first call within the same second.
Issue #85

A debounce timer is used to reduce the load on the Tuya device when multiple commands come in quick succession, which was added upstream to improve reliability.

The one second delay it adds to every unbatched command makes the local control even slower in some cases than using cloud control, so for the majority of cases where a single command is given in isolation, it should be avoided.

- Update tinytuya