Get entity category explicitly from config file.
Initially when I added this feature after 2021.11 HA release, I was looking
for a quick solution that did not involve mass editing all the existing
configs.  But in the course of other changes, I have done that anyway, and
added category attributes to secondary entities as I went.

Some secondary entities are just as important as the primary entity -
for example in a multi-switch powerboard, each switch is equal to the others,
and we shouldn't have one displaying as the main entity with others
demoted to config entities where Alexa/Google/Siri will not see them.