BAC002 thermostat: rearrange to make off/cool/heat/fan the primary controls.
The AC in the name seems to indicate that this is a full AC controller.
So switching between heat, cool and fan is not an installation option as
originally configured, but a primary function of the device.

To make way for these controls, program/manual selection was moved to a
separate select control.

Issue #84