Handle bitfields specially in mappings.
Until now, bitfields were just handled as integers, as we only ever
knew of one error code for devices that use them.

SmartMCB had documentation through the Tuya portal of the error field,
so we could add all the possible values, some of which could
conceivably happen concurrently, so it is time to treat bitfields
separately from integers where needed.