Inkbird ITC-306A: Add Unbalanced error binary sensor.
In case where the probes differ by more than 3C or 5F, the ITC-306A generates
an error code 32 (E4 on display), and asserts dp 113.
- Continuous heat error is actually asserted on dp 120, not 113. (E5 on display, error code 64).  The term "unbalanced" is used, as it does not involve introducing any new terminology in the translation file.

There are also E1 and E2 errors mentioned in the manual for probe fault. And there is no mention of the missing E3, but that would account for the third unknown boolean.  But since there appears to be no order to the existing dp assignment, we cannot guess as to which dp maps to which, and despite the explanation in the manual including "probe unplugged", they appear to be hard wired, so cannot be tested easily.