2022.4 compatibility fix.
In 2022.2, HA introduced Enumerations to replace strings in several locations.  After just two releases in 2022.4, the support for using strings was removed from categories, causing entities to fail to initialise if they specified a category in the config.

Switch to enums for EntityCategory, and update minimum supported HA version to 2022.2.

Temporarily disable automated testing, as the current version of test harness has a resource leak that causes tests to start failing after the 255th test, and we have over 1000 tests.  The last version without this problem required HA 2021.12 to run.