Fix condition for running SonarCloud scans
Since making this conditional, they have not been running at all.
It seems github.event_name is what was needed for testing for pushes.
Also, there needs to be a check that we are running on the main repository,
so that forks do not get failures due to unset SonarCloud credentials