Eliminate configuration of non-deprecated entities.
Home Assistant now has the ability to disable unwanted entities from the UI, and most non-essential entities are already marked as config or diagnostic so should not be cluttering up auto-generated dashboards or voice assistant entity lists.

Removing the ability to disable entities from the integration config simplifies the translations file, which is of much bigger benefit for maintenance than the benefit this feature brought.

Fixes #177, and a number of other comments that contributors have had about the same problem in the past.