Change condition for running SonarCloud scan.
Repository owner is me even when other users make a pull request, but contrary to SonarCloud docs, it does not use my repository's sonar token for pull requests.
Change the condition to basically check if the repository owner is the actor.
This allows others to set up SonarCloud on their forked repos if they want to
check their own code, but stops the errors from SonarCloud auth in pull requests.