Add support for Ledkia fan and light controller.
This has a light type switch according to the docs at iot.tuya.com, though the value is not showing in the sample data captured by the log.

Since the colour temperature and color options only make sense for LED lights, this is used to restrict those settings to only when LED lights are configured.  But if that value is never returned, then it will always only work as a dimmer.  In any case, we cannot know the capabilities of the LED bulb fitted - some are not even dimmable.

Issue #234