Poolex Q-line: use heat_cool rather than fan_only for "mute" mode.
Based on the comment "all are in grey", fan_only may be disabling the temperature setting, which would not be good.
Also, add a catch all default to avoid the reported error "Unrecognized HVAC Mode of True ignored".  Whether there is another missing mode, or there is just a period during startup when the mode is unavailble, this should avoid the issue.

Issue #166