Rename entities in accordance with HA guidelines.
In announcing the new naming API changes for 2022.8, there was also mention that only the first word in each device and entity name should be capitalised unless there are abbreviations or brand names that require it.

This is a bulk rename, for entities there should only be case changes to avoid changing entity ids, but some devices have been renamed (the filename is used for any unique id like usage which needs to be stable).