Don't try to invert non-numeric values.
Issue #241

Sometimes the value coming back from the device is None (for example many
lights do not return color_temp if they are not in color mode, but HA tries
to read the color_temp in various places if the light says it supports it).
In these cases, None should be returned from get_value(), but if the attribute
is configured to be inverted, an exception was thrown instead.