Prepare 0.19.3 release before starting more extensive changes
tinytuya 1.7.0 was released, and I want to update to that for protocol 3.2 and 3.4 support.  But after the 0.19.0 issues with tinytuya 1.6.6, this is likely to need extensive changes to the low level protocol handling so might take a while to stabilise.

In the meantime, current 0.19.2 release is incorrectly marked as 0.19.0 in the mainfest, and there are few minor device and documentation changes that are ready for release already.