Inventor dehumidifer: Allow non-integer target humidity
Another user trying to add their dehumidifier is getting failures because the
humidity dp is actually a string not an integer.  The values "40" and "50"
were documented, so it looked like it could be treated as an integer with
step 10 and range 40 to 50, but it seems the value "CO" (constant?) is also
a possible value.

Map "CO" to 30 (out of range so it cannot be manually selected - I suspect
"CO" is automatically used when in the previously undocumented "manual" mode).

Issue #263