Screen sync light: make switch optional, remove timer
Based on feedback on #264, switch is sometimes not returned, so must be optional to reliably detect the light.  Timer is also often not returned.  Since the entities are not designed to handle all attributes missing, move it to an attribute of the light (where it will be read-only when it is returned).