Tests: improve general config tests.
Add some checks for basic entity and dps requirements that aways need
to be met.

This is part of eliminating the need for writing device specific tests.
Lately I have given up on this because the tests are taking too long
to run, and the influx of new devices is too fast for writing tests.

Ideally there should be a simple declarative test config for each
device.  As each device becomes fully covered by that, the device
specific test classes can be removed, leaving only enough explicit
tests to maintain test coverage of the source code.

Also removed usage of TestCase.subTest, since pytest has a bug that
hides most of the output when it is used, and doesn't output the
context anyway.  Instead specify a message in each assert that
contains the context.