climate: overide turn_off and turn_on
Override the turn_off and turn_on functions for climate platform.
When the hvac_mode attribute is boolean, toggle it without considering
the mappings so that the previous mode is not lost when turning off
and back on again.

Caveat: a hypothetical Tuya device that uses True for off and False
for on will not work correctly since the mappings are bypassed.  Such
a device has never been encountered, and is unlikely to exist.

Modify old climate device configs which had coupled "power" and
"hvac_mode" dps to use "hvac_mode" and "mode" instead.  Newer configs
have done it this way, since the resulting config file is simpler, but
some older ones had the logic reversed due to old bugs that have since
been fixed.

This is quite a big change, with high risk of a typo messing up one or
more devices. Unit tests are passing, so any devices with tests should
be OK.  Mostly it is the newer devices that don't have tests, which
mostly did not need modifying, unless someone copied from an older
config in one of the PRs.

Issue #302