Add another generic smartplug variant to better match GNCC 16A one.
Issue #300

This smartplugv2 config is basically smartplugv2_energy with the energy sensor
removed (strangely it still seems to report the calibration for the energy
sensor though).

So it is more featured than the smartplugv1 and smartplugv2, but less featured
than smartplugv2_energy.

This config will not solve the bug reported that the switch does not work
from HA, but does reflect the state of the switch when switched on the
device itself or through the cloud.