Daizuki heatpump: make dp 101 and 125 optional.
These were captured in the first diagnostics captured from the device, and
since they are not in the starlight config appeared to be a way to differentiate
these two heatpumps.  However, later attempts were missing both, so they appear
to be optional (likely for the starlight too).  First attempt was in auto mode,
second in hot mode, so maybe these are limited to auto mode somehow.

Issue #276