Allow explicitly setting the protocol version.
With protocol versions 3.2 and 3.4 being similar to 3.3, there may be some misdetection happening, leading to issues such as reported in #300 where the device appears to work for returned values, but does not accept set values.

Add protocol version to the config, with default of "auto" that retains the same behaviour as previous versions.  When set to anything else, the protocol does not rotate to other versions when there is an error, so the user must know what version the device is using.