Remove test for matching translations between common items
Most of the items we wanted to check were removed.  Only name remains
in common between these two config_flow nodes, so the chances of mismatch
are much reduced.

Mainly this test was there to detect when a new item had been added but was only
added to one of the lists.