Invert the meaning of reversed.
It seems Tuya and Home Assistant have opposing views on what is normal
and what is reversed.  To make the behaviour consistent we need to
convert to Home Assistant's view, otherwise the reversed switch just
toggles between two different opposing views of the state Home Assistant
sees and the behaviour that Tuya has.