Add a third config for Beca BHP6000 thermostat quirks.
Issue #215

When set to C, the temperature in the protocol seems to start equal to
F at the bottom of the range, and go up by 0.5 for every 1 increment,
so by the top of the range has drifted away from F by about 4-5
degrees.

A third config has been added to handle this rather than replacing the
existing C config, as it is a bit strange and may be due to a firmware
bug in a specific version of the thermostat.