Fairy lights: separate scenes into a select entity.
Scenes require color_mode to be set to "scene", but if we add an "effect" dp,
extra values from color_mode are not selectable as effects.

Instead of putting scene as effect, separate it into its own select entity and
let the additional color_mode show through as effects.

Add the timer as a number entity instead of just an attribute.  It seems
an entity with only optional dps still works.

Issue #309