Refactor: hardcode default temperature unit in climate and water_heater entities.
Instead of deferring to device, just hardcode the fallback as CELSIUS.

I guess the original thinking was that temperature units would depend on the device (US marketed devices vs rest of world).  But if not specified explicitly in the config, we can only really have one default, which is CELSIUS.

This simplifies device by removing an unnecessary property, and removes that link with the history of this integration originally targeting climate devices.