Add tests for new functionality in device.py
The asynchronous loop is a bit hard to test.  It may need refactoring
into smaller, testable functions to improve the coverage further.

Don't register a listener for starting the loop later when Home
Assistant is shutting down.  Probably not important, as the listener
will not be notified, but seems cleaner.