Improve test coverage of device_config.py
Quite a few features built into the device_config logic do not get exercised by any current device configs, so test coverage is lacking.  This commit adds a few tests to cover all except some particularly complex code (which needs refactoring to reduce complexity, which should make it more easily tested).

Also removed the useless feature of being able to override the range for a dp from a mapping on the same dp (the ability to do so from a condition on a different dp is useful and remains).