Add a test for async_receive generator.
This should improve test coverage to above the quality threshold for recently
added code, as it will exercise most/all of the biggest function added.
The wrapper async task that calls this is still untested, but that is
relatively small. There are also a couple of optional listener callbacks
untested, as they have the side effect of starting that task.