Add Dual smartplug with power monitoring
Tuya portal also shows the add_ele for full energy monitoring, but it is not
showng up in the debug info from the user.  Anyway, these add_ele readings
have proven to be too transient to be useful since the devices seem to clear
them back to 0 after sending the increment to the portal.

Issue #342