Add support for Hysen HY08ACF thermostat.
This is much more limited in features than the HY08WE2 that is already
supported.  Since the diagrams show both hot and cold water pipes
being controlled by the device, but there is no selection at protocol
level, I assume that the "on" position is automatically selecting (or
even mixing) which pipe to open based on current temperature vs target
temperature, therefore the on position is set as heat_cool in Home Assistant.