light: modify handling of effect and color_mode
Based on testing, sending an explicit color_mode to switch modes is
not supported, color_mode is read-only in HA to reflect the status of
the light, and expected to be set implicitly depending on other
settings.

Previously, code to set the effect via color mode dp was unreachable,
since color mode was always initialised to the current color_mode,
so there was always an actual color_mode to be set. So rework
color_mode auto setting code to only set it if the current mode is not
suitable for the arguments supplied.

Update tests for the fact that color_mode cannot be explicitly set,
and implicit setting only sets it when needed.

Issue #333, #309