Get HA to call async_refresh once for each device during startup
Since the persistent connection changes, devices are appearing
unavailable until startup completes, since we have to wait for startup
to complete before opening any persistent connection in order to avoid
being detected as delaying startup ourselves.

If we can make one poll to get initial data, then the devices can
activate, even if something is delaying startup.  The risk though is
that unresponsive devices can then be responsible for delaying startup
as the connection will be active until they timeout.  But since we
only do this once per device, it should be less of a problem than it
was when we were polling.