Hysen HY08ACF: move child lock type to an attribute on child lock.
This does not seem to be editable as advertised in iot.tuya.com description.
Even in the Tuya app, changing it does nothing, and it just seems to be
reflecting a setting made from the device panel.  As it is not very useful,
demote it to an attribute on the related child lock entity.

Issue #343