motion sensor light: add an "Auto mode" switch
Add a switch for switching between auto and manual.  Since the effect
attribute is not available when the light is off, it is not so
convenient only having it there, because you can't tell the difference
between off and auto modes when the light is off.

Issue #282

To acheive this, the handling of conditions was modified to not try to
set the constraint dp when it is read-only, and instead require an
exact match.  This allows us to choose between "mode_off" and
"mode_on" when turning the auto mode switch off, to avoid changing the
light state as a side effect.  The previous condition handling was
selecting the first match for off and trying to change both dps to
make that condition match.

- as a side effect of these changes, the Fairland IPHCR15 heatpump also
changed behaviour, as it had a mode dp marked as readonly in the
config, used as conditions for hvac_mode.  A comment on #222 says
that despite being described as readonly in the Tuya portal, it
appeared possible to change it, so to avoid unexpected problems, the
readonly attribute was removed so that current behaviour is retained.