Use tinytuya set_multiple_values API function to send.
When this was originally developed, the only way to set arbitrary dps
was to work with low level functions, generating the payload then
sending it.

Over time, tinytuya has evolved with a more complete API, so there is
a higher level function to do what we want.

Currently it looks like it basically does the same as what we were
doing, but with new Tuya protocol versions coming all the time, and
workarounds for device issues being built in to tinytuya, it may pick
up extra functionality at some point, which we would miss by rolling
our own using lower level functions.

There is a minor performance hit in the case where we need resends,
but actually it looks like resends are handled at a lower level in
tinytuya anyway.

Unlikely based on review of the code, but may help with #338 and #372.