Refactor:  merge generic/* implementations into the top-level platforms.
In the early days, each device had its own implementation in Python.  After a small handful of devices, the move to using config files with a "generic" implementation was made so the classes did not need to keep multiplying.
But it took a while for some of the more complex logic in the early classes to be implemented in a way that could be configured using conditional mappings etc, so initially the basic generic implementations were alongside the old device specific implementations, with the top level platforms just doing loading of HA config and delegating to the various implementations.

For some time now, we have been rid of the device specific implementations, and just have one generic implementation for each platform. Since there is a one to one mapping, it is easier to to follow the code if the generic implementation is merged with the top level platform file.